feat(review): Renames classes and move files
